feat(Title): move agent Title component to Title/Agent/ folder
