+++ slash/css/src/Title/Agent/Title.scss
@@ -1,5 +1,5 @@
 @use "sass:color";
-@use "../common/common.agent.scss" as common;
+@use "../../common/common.agent.scss" as common;
 
 h2.af-title {
   font-size: 1.5rem;
