+++ slash/css/src/agent.scss
@@ -11,7 +11,7 @@
 @use "./Form/Text/Agent/Text.agent";
 @use "./Form/Textarea/Textarea.agent";
 @use "./Modal/Modal.agent";
-@use "./Title/Title.agent";
+@use "./Title/Agent/Title";
 @use "./Layout/Footer/Footer.agent";
 @use "./Badge/Badge";
 @use "./Tabs/Tabs.agent";
