+++ slash/react/src/Title/Agent/Title.mdx
@@ -1,5 +1,5 @@
 import { Controls, Canvas, Meta } from "@storybook/blocks";
-import * as TitleStories from "./Title.agent.stories";
+import * as TitleStories from "./Title.stories";
 
 <Meta of={TitleStories} />
